requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.3.0
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
limits==5.5.0
packaging==25.0
passlib==1.7.4
prometheus-client==0.19.0
pyasn1==0.6.1
pydantic==2.5.0
pydantic_core==2.14.1
PyJWT==2.8.0
pymongo==4.6.0
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.6
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.14.1
uvicorn==0.24.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3
